+++ CChromaEditorLibrary/EditorAnimation2D.cpp
@@ -73,7 +73,7 @@ void EditorAnimation2D::OverrideTime(float time)
 float EditorAnimation2D::GetDuration(int index)
 {
 	vector<FChromaSDKColorFrame2D>& frames = GetFrames();
-	if (index < frames.size())
+	if (index < (int)frames.size())
 	{
 		FChromaSDKColorFrame2D& frame = frames[index];
 		return frame.Duration;
@@ -120,11 +120,11 @@ void EditorAnimation2D::CopyPixels(COLORREF* pColor, const UINT width, const UIN
 	vector<FChromaSDKColorFrame2D>& frames = GetFrames();
 	int currentFrame = GetCurrentFrame();
 	if (currentFrame < 0 ||
-		currentFrame >= frames.size())
+		currentFrame >= (int)frames.size())
 	{
 		currentFrame = 0;
 	}
-	if (currentFrame < frames.size())
+	if (currentFrame < (int)frames.size())
 	{
 		frames[currentFrame] = GetCopy();
 	}
